+++ src/main/java/com/ktb/interview/controller/InterviewSessionController.java
@@ -12,8 +12,8 @@ import com.ktb.interview.dto.response.InterviewRealSubmitResponse;
 import com.ktb.interview.dto.response.InterviewSessionStateResponse;
 import com.ktb.interview.dto.response.SessionFeedbackFailedResponse;
 import com.ktb.interview.dto.response.SessionFeedbackPendingResponse;
-import com.ktb.interview.application.InterviewSessionManagementService;
-import com.ktb.interview.application.InterviewSubmissionService;
+import com.ktb.interview.service.InterviewSessionManagementService;
+import com.ktb.interview.service.InterviewSubmissionService;
 import com.ktb.interview.domain.InterviewSessionStatus;
 import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.auth.security.adapter.SecurityUserAccount;
