+++ src/main/java/com/ktb/interview/service/impl/InterviewSessionIdProvider.java
@@ -1,0 +1,18 @@
+package com.ktb.interview.service.impl;
+
+import java.util.UUID;
+import org.springframework.stereotype.Component;
+
+/**
+ * 인터뷰 세션 식별자 생성기.
+ */
+@Component
+public class InterviewSessionIdProvider {
+
+    /**
+     * 외부 노출 가능한 무작위 세션 ID를 생성합니다.
+     */
+    public String nextId() {
+        return UUID.randomUUID().toString();
+    }
+}
