+++ src/main/java/com/ktb/interview/service/impl/InterviewSessionManagementServiceImpl.java
@@ -1,0 +1,49 @@
+package com.ktb.interview.service.impl;
+
+import com.ktb.interview.service.InterviewSessionManagementService;
+import com.ktb.interview.service.impl.flow.InterviewSessionCreateFlowService;
+import com.ktb.interview.service.impl.flow.InterviewSessionFeedbackQueryFlowService;
+import com.ktb.interview.dto.request.InterviewSessionCreateRequest;
+import com.ktb.interview.dto.response.InterviewSessionCreateResponse;
+import com.ktb.interview.dto.response.InterviewSessionFinalFeedbackResponse;
+import com.ktb.interview.dto.response.InterviewSessionStateResponse;
+import com.ktb.interview.dto.response.SessionFeedbackFailedResponse;
+import com.ktb.interview.dto.response.SessionFeedbackPendingResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+/**
+ * 인터뷰 세션 관리 유스케이스의 진입점 파사드.
+ */
+@Service
+@RequiredArgsConstructor
+public class InterviewSessionManagementServiceImpl implements InterviewSessionManagementService {
+
+    private final InterviewSessionCreateFlowService sessionCreateFlowService;
+    private final InterviewSessionFeedbackQueryFlowService sessionFeedbackQueryFlowService;
+
+    @Override
+    public InterviewSessionCreateResponse createSession(Long accountId, InterviewSessionCreateRequest request) {
+        return sessionCreateFlowService.createSession(accountId, request);
+    }
+
+    @Override
+    public InterviewSessionStateResponse getSessionState(Long accountId, String sessionId) {
+        return sessionFeedbackQueryFlowService.getSessionState(accountId, sessionId);
+    }
+
+    @Override
+    public SessionFeedbackPendingResponse getSessionFeedbackPending(Long accountId, String sessionId) {
+        return sessionFeedbackQueryFlowService.getSessionFeedbackPending(accountId, sessionId);
+    }
+
+    @Override
+    public SessionFeedbackFailedResponse getSessionFeedbackFailed(Long accountId, String sessionId) {
+        return sessionFeedbackQueryFlowService.getSessionFeedbackFailed(accountId, sessionId);
+    }
+
+    @Override
+    public InterviewSessionFinalFeedbackResponse getSessionFeedbackCompleted(Long accountId, String sessionId) {
+        return sessionFeedbackQueryFlowService.getSessionFeedbackCompleted(accountId, sessionId);
+    }
+}
